broken into components, change name components, fix NavLink active
