--- src/ProfilePage/Header/Header.js
@@ -1,12 +1,0 @@
-import React from "react";
-import TopNavbar from "./TopNavbar";
-import ProfileNavigation from "./ProfileNavbar";
-
-export default () => {
-  return (
-    <div>
-      <TopNavbar />
-      <ProfileNavigation />
-    </div>
-  );
-};
