+++ src/ProfilePage/Header/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Navigation from "./Navigation";
+import Statistics from "./Statistics";
+
+export default () => {
+  return (
+    <div>
+      <Navigation />
+      <Statistics />
+    </div>
+  );
+};
